requirements.txt
FastAPI>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
pyswisseph==2.10.3.2
